hide add group headings if there are no records.

+++ core/mvc/z/admin/controller/site-options.cfc
@@ -3901,7 +3901,7 @@ Define this function in another CFC to override the default email format
 		
 		rowStruct={};
 		rowIndexFix=1;
-		if(structkeyexists(arguments.struct, 'recurse')){
+		if(structkeyexists(arguments.struct, 'recurse') and qS.recordcount NEQ 0){
 			echo('<h3>Sub-group: #q12.site_option_group_display_name#</h3>');
 		}
 		addEnabled=true;
@@ -3911,14 +3911,16 @@ Define this function in another CFC to override the default email format
 					addEnabled=false;
 				}
 			}
-			echo('<p>');
-			if(addEnabled){
-				writeoutput('<a href="#application.zcore.functions.zURLAppend(arguments.struct.addURL, "site_option_app_id=#form.site_option_app_id#&amp;site_option_group_id=#form.site_option_group_id#&amp;site_x_option_group_set_parent_id=#form.site_x_option_group_set_parent_id#")#">Add #htmleditformat(application.zcore.functions.zFirstLetterCaps(qGroup.site_option_group_display_name))#</a>');
-			} 
-			if(methodBackup EQ "manageGroup"){
-				echo(' | <a href="/z/admin/site-option-group/export?site_option_group_id=#qGroup.site_option_group_id#" target="_blank">Export CSV</a>');
+			if(structkeyexists(arguments.struct, 'recurse') EQ false or qS.recordcount NEQ 0){
+				echo('<p>');
+				if(addEnabled){
+					writeoutput('<a href="#application.zcore.functions.zURLAppend(arguments.struct.addURL, "site_option_app_id=#form.site_option_app_id#&amp;site_option_group_id=#form.site_option_group_id#&amp;site_x_option_group_set_parent_id=#form.site_x_option_group_set_parent_id#")#">Add #htmleditformat(application.zcore.functions.zFirstLetterCaps(qGroup.site_option_group_display_name))#</a>');
+				} 
+				if(methodBackup EQ "manageGroup"){
+					echo(' | <a href="/z/admin/site-option-group/export?site_option_group_id=#qGroup.site_option_group_id#" target="_blank">Export CSV</a>');
+				}
+				echo('</p>');
 			}
-			echo('</p>');
 		} 
 		if(not structkeyexists(arguments.struct, 'recurse')){
 			if(qGroup.site_option_group_enable_sorting EQ 1 and subgroupRecurseEnabled){
